requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
passlib[bcrypt]==4.0.1
python-jose[cryptography]
pydantic-settings
pdfkit
python-multipart
